move the rust functionality to deptry.rust
