+++ python/deptry/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+import logging
+
+logging.getLogger("nbconvert").setLevel(logging.WARNING)
