+++ python/deptry/dependency.py
