+++ python/deptry/reporters/json.py
