+++ python/deptry/reporters/text.py
