+++ python/deptry/rust.pyi
@@ -1,4 +1,4 @@
-from deptry import Location as RustLocation
+from .rust import Location as RustLocation
 
 def get_imports_from_py_files(file_paths: list[str]) -> dict[str, list[RustLocation]]: ...
 def get_imports_from_py_file(file_path: str) -> dict[str, list[RustLocation]]: ...
